перенес данные FILMS в директорию для моков
