+++ project/src/components/film-card/film-card.tsx
@@ -1,22 +1,4 @@
-type Film = {
-  name: string,
-  posterImage: string,
-  previewImage: string,
-  backgroundImage: string,
-  backgroundColor: string,
-  description: string,
-  rating: number,
-  scoresCount: number,
-  director: string,
-  starring: string[],
-  runTime: number,
-  genre: string,
-  released: number,
-  id: number,
-  isFavorite: boolean,
-  videoLink: string,
-  previewVideoLink: string,
-}
+import { Film } from '../../mocks/films';
 
 type FilmCardProps = {
   film: Film;
