+++ project/src/index.tsx
@@ -1,515 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './components/app/app';
-
-type Film = {
-  name: string,
-  posterImage: string,
-  previewImage: string,
-  backgroundImage: string,
-  backgroundColor: string,
-  description: string,
-  rating: number,
-  scoresCount: number,
-  director: string,
-  starring: string[],
-  runTime: number,
-  genre: string,
-  released: number,
-  id: number,
-  isFavorite: boolean,
-  videoLink: string,
-  previewVideoLink: string,
-}
-
-const PROMO_FILM: Film =
-{
-  name: 'The Revenant',
-  posterImage: 'https://9.react.pages.academy/static/film/poster/Revenant.jpg',
-  previewImage: 'https://9.react.pages.academy/static/film/preview/revenant.jpg',
-  backgroundImage: 'https://9.react.pages.academy/static/film/background/Revenant.jpg',
-  backgroundColor: '#92918B',
-  description: 'A frontiersman on a fur trading expedition in the 1820s fights for survival after being mauled by a bear and left for dead by members of his own hunting team.',
-  rating: 4,
-  scoresCount: 618498,
-  director: 'Alejandro G. Iñárritu',
-  starring: [
-    'Leonardo DiCaprio',
-    'Tom Hardy',
-    'Will Poulter',
-  ],
-  runTime: 156,
-  genre: 'Action',
-  released: 2015,
-  id: 1,
-  isFavorite: false,
-  videoLink: 'https://9.react.pages.academy/static/film/video/bubbles.mp4',
-  previewVideoLink: 'https://9.react.pages.academy/static/film/video/dog.mp4',
-};
-
-const FILMS: Film[] =
-[
-  {
-    name: 'The Revenant',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/Revenant.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/revenant.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/Revenant.jpg',
-    backgroundColor: '#92918B',
-    description: 'A frontiersman on a fur trading expedition in the 1820s fights for survival after being mauled by a bear and left for dead by members of his own hunting team.',
-    rating: 4,
-    scoresCount: 618498,
-    director: 'Alejandro G. Iñárritu',
-    starring: [
-      'Leonardo DiCaprio',
-      'Tom Hardy',
-      'Will Poulter',
-    ],
-    runTime: 156,
-    genre: 'Action',
-    released: 2015,
-    id: 1,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bubbles.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/dog.mp4',
-  },
-  {
-    name: 'Orlando',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/Orlando.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/orlando.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/Orlando.jpg',
-    backgroundColor: '#D8D3BD',
-    description: 'Young nobleman Orlando is commanded by Queen Elizabeth I to stay forever young. Miraculously, he does just that. The film follows him as he moves through several centuries of British history, experiencing a variety of lives and relationships along the way, and even changing sex.',
-    rating: 2.6,
-    scoresCount: 12292,
-    director: 'Sally Potter',
-    starring: [
-      'Tilda Swinton',
-      'Billy Zane',
-      'Quentin Crisp',
-    ],
-    runTime: 94,
-    genre: 'Drama',
-    released: 1992,
-    id: 2,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bike.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/traffic.mp4',
-  },
-  {
-    name: 'We need to talk about Kevin',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/We_need_to_talk_about_Kevin.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/we-need-to-talk-about-kevin.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/We_need_to_talk_about_Kevin.jpg',
-    backgroundColor: '#E1DFDE',
-    description: 'Kevin \'s mother struggles to love her strange child, despite the increasingly dangerous things he says and does as he grows up. But Kevin is just getting started, and his final act will be beyond anything anyone imagined.',
-    rating: 3.8,
-    scoresCount: 123240,
-    director: 'Lynne Ramsay',
-    starring: [
-      'Tilda Swinton',
-      'John C. Reilly',
-      'Ezra Miller',
-    ],
-    runTime: 112,
-    genre: 'Drama',
-    released: 2011,
-    id: 3,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bubbles.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/traffic.mp4',
-  },
-  {
-    name: 'Seven Years in Tibet',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/Seven_Years_in_Tibet.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/seven-years-in-tibet.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/Seven_Years_in_Tibet.jpg',
-    backgroundColor: '#C6CADF',
-    description: 'True story of Heinrich Harrer, an Austrian mountain climber who became friends with the Dalai Lama at the time of China takeover of Tibet.',
-    rating: 3.6,
-    scoresCount: 112612,
-    director: 'Jean-Jacques Annaud',
-    starring: [
-      'Brad Pitt',
-      'David Thewlis',
-      'BD Wong',
-    ],
-    runTime: 136,
-    genre: 'Adventure',
-    released: 1997,
-    id: 4,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bubbles.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/dog.mp4',
-  },
-  {
-    name: 'Fantastic Beasts: The Crimes of Grindelwald',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/Fantastic_Beasts.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/fantastic-beasts-the-crimes-of-grindelwald.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/Fantastic_Beasts.jpg',
-    backgroundColor: '#B6A99F',
-    description: 'In an effort to thwart Grindelwald\'s plans of raising pure-blood wizards to rule over all non-magical beings, Albus Dumbledore enlists his former student Newt Scamander, who agrees to help, though he\'s unaware of the dangers that lie ahead. Lines are drawn as love and loyalty are tested, even among the truest friends and family, in an increasingly divided wizarding world.',
-    rating: 3.4,
-    scoresCount: 160757,
-    director: 'David Yates',
-    starring: [
-      'Eddie Redmayne',
-      'Katherine Waterston',
-      'Dan Fogler',
-    ],
-    runTime: 134,
-    genre: 'Fantasy',
-    released: 2018,
-    id: 5,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bubbles.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/dog.mp4',
-  },
-  {
-    name: 'Macbeth',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/Macbeth.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/macbeth.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/Macbeth.jpg',
-    backgroundColor: '#F1E9CE',
-    description: 'Macbeth, the Thane of Glamis, receives a prophecy from a trio of witches that one day he will become King of Scotland. Consumed by ambition and spurred to action by his wife, Macbeth murders his king and takes the throne for himself.',
-    rating: 3.3,
-    scoresCount: 48798,
-    director: 'Justin Kurzel',
-    starring: [
-      'Michael Fassbender',
-      'Marion Cotillard',
-      'Jack Madigan',
-    ],
-    runTime: 113,
-    genre: 'Drama',
-    released: 2015,
-    id: 6,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bike.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/traffic.mp4',
-  },
-  {
-    name: 'Matrix',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/matrix.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/matrix.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/matrix.jpg',
-    backgroundColor: '#B9B27E',
-    description: 'A computer hacker learns from mysterious rebels about the true nature of his reality and his role in the war against its controllers.',
-    rating: 4.4,
-    scoresCount: 1503092,
-    director: 'Wachowski Brothers',
-    starring: [
-      'Keanu Reeves',
-      'Laurence Fishburne',
-      'Carrie-Anne Moss',
-    ],
-    runTime: 136,
-    genre: 'Action',
-    released: 1999,
-    id: 7,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bike.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/traffic.mp4',
-  },
-  {
-    name: 'No Country for Old Men',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/No_Country_for_Old_Men.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/no-country-for-old-men.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/No_Country_for_Old_Men.jpg',
-    backgroundColor: '#BDAD8F',
-    description: 'Violence and mayhem ensue after a hunter stumbles upon a drug deal gone wrong and more than two million dollars in cash near the Rio Grande.',
-    rating: 4.1,
-    scoresCount: 764976,
-    director: 'Ethan Coen',
-    starring: [
-      'Tommy Lee Jones',
-      'Javier Bardem',
-      'Josh Brolin',
-    ],
-    runTime: 122,
-    genre: 'Crime',
-    released: 2007,
-    id: 8,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bubbles.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/traffic.mp4',
-  },
-  {
-    name: 'Bronson',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/bronson.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/bronson.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/bronson.jpg',
-    backgroundColor: '#73B39A',
-    description: 'A young man who was sentenced to seven years in prison for robbing a post office ends up spending three decades in solitary confinement. During this time, his own personality is supplanted by his alter-ego, Charles Bronson.',
-    rating: 3.6,
-    scoresCount: 109661,
-    director: 'Nicolas Winding Refn',
-    starring: [
-      'Tom Hardy',
-      'Kelly Adams',
-      'Luing Andrews',
-    ],
-    runTime: 92,
-    genre: 'Action',
-    released: 2008,
-    id: 9,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bubbles.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/dog.mp4',
-  },
-  {
-    name: 'Bohemian Rhapsody',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/Bohemian_Rhapsody.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/bohemian_rhapsody.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/Bohemian_Rhapsody.jpg',
-    backgroundColor: '#929FA5',
-    description: 'Bohemian Rhapsody is a foot-stomping celebration of Queen, their music and their extraordinary lead singer Freddie Mercury. Freddie defied stereotypes and shattered convention to become one of the most beloved entertainers on the planet. The film traces the meteoric rise of the band through their iconic songs and revolutionary sound. They reach unparalleled success, but in an unexpected turn Freddie, surrounded by darker influences, shuns Queen in pursuit of his solo career. Having suffered greatly without the collaboration of Queen, Freddie manages to reunite with his bandmates just in time for Live Aid. While bravely facing a recent AIDS diagnosis, Freddie leads the band in one of the greatest performances in the history of rock music. Queen cements a legacy that continues to inspire outsiders, dreamers and music lovers to this day.',
-    rating: 6.1,
-    scoresCount: 338903,
-    director: 'Bryan Singer',
-    starring: [
-      'Rami Malek',
-      'Lucy Boynton',
-      'Gwilym Lee',
-    ],
-    runTime: 134,
-    genre: 'Drama',
-    released: 2018,
-    id: 10,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/matrix.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/dog.mp4',
-  },
-  {
-    name: 'Moonrise Kingdom',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/Moonrise_Kingdom.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/moonrise-kingdom.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/Moonrise_Kingdom.jpg',
-    backgroundColor: '#D8E3E5',
-    description: 'A pair of young lovers flee their New England town, which causes a local search party to fan out to find them.',
-    rating: 7.9,
-    scoresCount: 291183,
-    director: 'Wes Anderson',
-    starring: [
-      'Jared Gilman',
-      'Kara Hayward',
-      'Bruce Willis',
-    ],
-    runTime: 94,
-    genre: 'Adventure',
-    released: 2012,
-    id: 11,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bike.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/traffic.mp4',
-  },
-  {
-    name: 'Once Upon a Time in America',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/Once_Upon_a_Time_in_America.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/Once_Upon_a_Time_in_America.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/ones_upon_a_time_in_america.jpg',
-    backgroundColor: '#CBAC79',
-    description: 'A former Prohibition-era Jewish gangster returns to the Lower East Side of Manhattan over thirty years later, where he once again must confront the ghosts and regrets of his old life.',
-    rating: 9.9,
-    scoresCount: 276395,
-    director: 'Sergio Leone',
-    starring: [
-      'Robert De Niro',
-      'James Woods',
-      'Elizabeth McGovern',
-    ],
-    runTime: 229,
-    genre: 'Crime',
-    released: 1984,
-    id: 12,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bike.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/dog.mp4',
-  },
-  {
-    name: 'Shutter Island',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/Shutter_Island.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/shutter-island.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/Shutter_Island.jpg',
-    backgroundColor: '#977461',
-    description: 'In 1954, a U.S. Marshal investigates the disappearance of a murderer, who escaped from a hospital for the criminally insane.',
-    rating: 4.1,
-    scoresCount: 1002557,
-    director: 'Martin Scorsese',
-    starring: [
-      'Leonardo DiCaprio',
-      'Emily Mortimer',
-      'Mark Ruffalo',
-    ],
-    runTime: 138,
-    genre: 'Thriller',
-    released: 2010,
-    id: 13,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bike.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/traffic.mp4',
-  },
-  {
-    name: 'What We Do in the Shadows',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/What-We-Do-in-the-Shadows.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/what-we-do-in-the-shadows.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/What-We-Do-in-the-Shadows.jpg',
-    backgroundColor: '#A39E81',
-    description: 'A look into the daily (or rather, nightly) lives of three vampires who\'ve lived together for over 100 years, in Staten Island.',
-    rating: 7.2,
-    scoresCount: 6173,
-    director: 'Jemaine Clement',
-    starring: [
-      'Kayvan Novak',
-      'Matt Berry',
-      'Natasia Demetriou',
-    ],
-    runTime: 30,
-    genre: 'Comedy',
-    released: 2019,
-    id: 14,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bike.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/dog.mp4',
-  },
-  {
-    name: 'Dardjeeling Limited',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/Dardjeeling_Limited.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/dardjeeling_limited.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/Dardjeeling_Limited.jpg',
-    backgroundColor: '#AD9F8B',
-    description: 'A year after their father\'s funeral, three brothers travel across India by train in an attempt to bond with each other.',
-    rating: 3.6,
-    scoresCount: 165106,
-    director: 'Wes Anderson',
-    starring: [
-      'Owen Wilson',
-      'Adrien Brody',
-      'Jason Schwartzman',
-    ],
-    runTime: 91,
-    genre: 'Adventure',
-    released: 2007,
-    id: 15,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bubbles.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/traffic.mp4',
-  },
-  {
-    name: 'War of the Worlds',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/War_of_the_Worlds.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/war-of-the-worlds.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/War_of_the_Worlds.jpg',
-    backgroundColor: '#9B7E61',
-    description: 'As Earth is invaded by alien tripod fighting machines, one family fights for survival.',
-    rating: 9.3,
-    scoresCount: 386834,
-    director: 'Steven Spielberg',
-    starring: [
-      'Tom Cruise',
-      'Dakota Fanning',
-      'Tim Robbins',
-    ],
-    runTime: 116,
-    genre: 'Adventure',
-    released: 2005,
-    id: 16,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/matrix.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/traffic.mp4',
-  },
-  {
-    name: 'Midnight Special',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/Midnight_Special.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/midnight-special.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/Midnight_Special.jpg',
-    backgroundColor: '#828585',
-    description: 'A father and son go on the run, pursued by the government and a cult drawn to the child\'s special powers.',
-    rating: 3.3,
-    scoresCount: 67815,
-    director: 'Jeff Nichols',
-    starring: [
-      'Michael Shannon',
-      'Joel Edgerton',
-      'Kirsten Dunst',
-    ],
-    runTime: 112,
-    genre: 'Action',
-    released: 2016,
-    id: 17,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bubbles.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/dog.mp4',
-  },
-  {
-    name: 'Snatch',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/Snatch.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/snatch.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/Snatch.jpg',
-    backgroundColor: '#FDFDFC',
-    description: 'Unscrupulous boxing promoters, violent bookmakers, a Russian gangster, incompetent amateur robbers and supposedly Jewish jewelers fight to track down a priceless stolen diamond.',
-    rating: 0.2,
-    scoresCount: 716577,
-    director: 'Guy Ritchie',
-    starring: [
-      'Jason Statham',
-      'Brad Pitt',
-      'Benicio Del Toro',
-    ],
-    runTime: 104,
-    genre: 'Comedy',
-    released: 2000,
-    id: 18,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bubbles.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/traffic.mp4',
-  },
-  {
-    name: 'Pulp Fiction',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/Pulp_Fiction.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/pulp-fiction.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/Pulp_Fiction.jpg',
-    backgroundColor: '#795433',
-    description: 'The lives of two mob hitmen, a boxer, a gangster & his wife, and a pair of diner bandits intertwine in four tales of violence and redemption.',
-    rating: 1.5,
-    scoresCount: 1635992,
-    director: 'Quentin Tarantino',
-    starring: [
-      'John Travolta',
-      'Uma Thurman',
-      'Samuel L. Jackson',
-    ],
-    runTime: 153,
-    genre: 'Crime',
-    released: 1994,
-    id: 19,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/bike.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/dog.mp4',
-  },
-  {
-    name: 'Legend',
-    posterImage: 'https://9.react.pages.academy/static/film/poster/Legend.jpg',
-    previewImage: 'https://9.react.pages.academy/static/film/preview/legend.jpg',
-    backgroundImage: 'https://9.react.pages.academy/static/film/background/legend.jpg',
-    backgroundColor: '#E1DAD7',
-    description: 'Identical twin gangsters Ronald and Reginald Kray terrorize London during the 1960s.',
-    rating: 3.5,
-    scoresCount: 138487,
-    director: 'Brian Helgeland',
-    starring: [
-      'Tom Hardy',
-      'Emily Browning',
-      'Taron Egerton',
-    ],
-    runTime: 132,
-    genre: 'Crime',
-    released: 2015,
-    id: 20,
-    isFavorite: false,
-    videoLink: 'https://9.react.pages.academy/static/film/video/matrix.mp4',
-    previewVideoLink: 'https://9.react.pages.academy/static/film/video/traffic.mp4',
-  },
-];
+import { FILMS, PROMO_FILM } from './mocks/films';
 
 ReactDOM.render(
   <React.StrictMode>
